Version 2 without conditional continue
This patchset changes the define to accept a conditional. The conditional is the true or false that was lifted from the if statements. We can combine both the conditional and the debug statement since the debug statement will always be true -- except it ends up being false due to the mis-compilation. So at the C level the logic is sound and can be verified.

Signed-off-by: Scotty Bauer <sbauer@eng.utah.edu>